make toplevel vr use flatlist; move view works button there
